Reworking the auto-submit feature one more time. We now have a lightweight submit that only queries parameters (without doing pagination) and a normal parameter submit.
On init, we always fire the lightweight submit, validate the parameters and if then autosubmit is reported to be enabled we fire the pagination request. Therefore we ensure that we do heavy work only if autosubmit is enabled.

The pagination controls act as auxillary submit buttons - so browsing through the pages submits the parameter content and paginates the report (same behaviour as before).