[CLEANUP] Separated IAsnycReportState from IAsyncReportListener interface.
Introduced explicit data object for AsyncReportState. This allows us to keep the status-listener clean and hides the implementation detail that our current implementation is also responsible for creating the state object.
Moved creation of the listener into the PentahoAsyncReportExecution class, as details on how this implementation is used here should not be part of the public interface. The executor now provides the callable with an UUID when it is scheduled, but does not manipulate or provide the listener implementation itself. Our public interface states that the reportexecutor is responsible for delivering a state. The fact that our listener implementation produces the state is now hidden from the interface.