[BACKLOG-7598] - Fix chrome pdf viewver plugin bogus behaviour
This brings next fixes for us:
1) We need a special endpoint for pdf viewer plugin now able to send get
request to whatever it want it to send. Since we can't fix chrome, we have
to fix JobManager fot that.

2) We need to keep same content between requests.
So since we expected to keep content we don't clean-up future after
content requested. Instead - we keep a collection of futures. We expect
clean-up when session destroyed. Also content is kept in staging folder.

3) Since we keep content - we introduce new methods for
IFixedSizeStreamingContent.cleanContent() to explicitly clean up content.

And yes - finally we came to a one-more cahce layer - all content is kept in
statging folder during server works.

Pay attention - content for all currently running sessions.